requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
python-dotenv==1.0.1

# OCR basics
pytesseract==0.3.13
Pillow==10.4.0
# Pillow's ImageGrab on Windows uses pywin32 indirectly via OS APIs; include pywin32 for robustness
pywin32==306

# Computer vision / numerical
opencv-python==4.10.0.84
# Use NumPy 1.26.x to satisfy faiss-cpu (<2.0) while remaining compatible with Python 3.11 and OpenCV
numpy==1.26.4

# Scheduling
APScheduler==3.10.4

# Hugging Face & transformers
huggingface_hub==0.26.2
transformers==4.44.2
sentence-transformers==3.1.1

# Vector search (choose FAISS CPU for now)
faiss-cpu==1.8.0.post1

# Orchestration (optional, but included for convenience)
ray==2.37.0

# HTTP / parsing utilities for knowledge ingestion
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Tooling
pre-commit==3.8.0
ruff==0.6.9
black==24.8.0
mypy==1.11.2
